backend/requirements.txt
uvicorn==0.18.2
fastapi==0.78.0
SQLAlchemy==1.4.29
psycopg2-binary==2.9.3
pydantic==1.9.1
python-multipart==0.0.5
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.0
PyYAML==6.0
ua-parser==0.16.1
user-agents==2.2.0
pymediainfo==5.1.0
ImageHash==4.2.1
Pillow==8.4.0
bitstring==3.1.9